cargo-release-plan: cover proposal diagnostics and organize tests
Exercise verbose and artifact failure paths, and split proposal scenarios into focused modules with shared test fixtures.
