+++ packages/cargo-release-plan/src/propose/tests/mod.rs
@@ -1,0 +1,9 @@
+pub(crate) use fixture::*;
+pub(crate) use harness::*;
+
+mod alignment;
+mod fixture;
+mod generation;
+mod harness;
+mod invariants;
+mod propagation;
